+++ src/beat/mod.rs
@@ -57,6 +57,7 @@ where
     default_queue: String,
     task_routes: Vec<(String, String)>,
     task_options: TaskOptions,
+    max_sleep_duration: Option<Duration>,
 }
 
 /// Used to create a [`Beat`] app with a custom configuration.
@@ -87,6 +88,7 @@ where
                 default_queue: "celery".into(),
                 task_routes: vec![],
                 task_options: TaskOptions::default(),
+                max_sleep_duration: None,
             },
             scheduler_backend: LocalSchedulerBackend::new(),
         }
@@ -116,6 +118,7 @@ where
                 default_queue: "celery".into(),
                 task_routes: vec![],
                 task_options: TaskOptions::default(),
+                max_sleep_duration: None,
             },
             scheduler_backend,
         }
@@ -170,6 +173,14 @@ where
         self
     }
 
+    /// Set a maximum sleep duration, which limits the amount of time that
+    /// can pass between ticks. This is useful to ensure that the scheduler backend
+    /// implementation is called regularly.
+    pub fn max_sleep_duration(mut self, max_sleep_duration: Duration) -> Self {
+        self.config.max_sleep_duration = Some(max_sleep_duration);
+        self
+    }
+
     /// Construct a `Beat` app with the current configuration.
     pub async fn build(self) -> Result<Beat<Bb::Broker, Sb>, CeleryError> {
         // Declare default queue to broker.
@@ -206,6 +217,7 @@ where
             broker_connection_retry: self.config.broker_connection_retry,
             broker_connection_max_retries: self.config.broker_connection_max_retries,
             broker_connection_retry_delay: self.config.broker_connection_retry_delay,
+            max_sleep_duration: self.config.max_sleep_duration,
         })
     }
 }
@@ -228,6 +240,8 @@ pub struct Beat<Br: Broker, Sb: SchedulerBackend> {
     broker_connection_retry: bool,
     broker_connection_max_retries: u32,
     broker_connection_retry_delay: u32,
+
+    max_sleep_duration: Option<Duration>,
 }
 
 impl<Br> Beat<Br, LocalSchedulerBackend>
@@ -360,6 +374,10 @@ where
                 let sleep_interval = next_tick_at.duration_since(now).expect(
                     "Unexpected error when unwrapping a SystemTime comparison that is not supposed to fail",
                 );
+                let sleep_interval = match &self.max_sleep_duration {
+                    Some(max_sleep_duration) => std::cmp::min(sleep_interval, *max_sleep_duration),
+                    None => sleep_interval,
+                };
                 debug!("Now sleeping for {:?}", sleep_interval);
                 time::sleep(sleep_interval).await;
             }