+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Redis broker support.
+- Added the `max_sleep_duration` property on the `Beat` which can be used to ensure that
+  the scheduler backend is called regularly (which may be necessary for custom backends).
 
 ## [v0.4.0-rc5](https://github.com/rusty-celery/rusty-celery/releases/tag/v0.4.0-rc5) - 2020-11-19